additional changes

+++ backend/controllers/userController.js
@@ -115,9 +115,6 @@ const getUserProfile = asyncHandler(async (req, res) => {
 const updateUserProfile = asyncHandler(async (req, res) => {
   const user = await User.findById(req.user._id);
 
-  console.log(user);
-
-
   if(user){
     user.name = req.body.name || user.name; // Check if the request came with a name, if not keep the name
     user.email = req.body.email || user.email; // Check if the request came with an email, if not keep the one the user already have
@@ -177,6 +174,7 @@ const updateUser = asyncHandler(async (req, res) => {
   if(user){
     user.name = req.body.name || user.name; // Check if the request came with a name, if not keep the name
     user.email = req.body.email || user.email; // Check if the request came with an email, if not keep the one the user already have
+    user.isAdmin = Boolean(req.body.isAdmin)
 
     if(req.body.password){ // Because the password is hashed, we cannot do it the way we did it above. Othwerwise we would need to dehash it first and then compare the 2
 
@@ -201,14 +199,22 @@ const updateUser = asyncHandler(async (req, res) => {
 });
 
 // @desc Delete user
-// @route DELETE /api/users/
+// @route DELETE /api/users/:id
 // @access Private
 const deleteUser = asyncHandler(async (req, res) => {
 
   const user = await User.findById(req.params.id)
+
   if(user){
+
+    if (user.isAdmin) {
+      res.status(400)
+      throw new Error('Cannot delete admin user')
+    }
+    
     await User.deleteOne({_id: user._id})
     res.status(200).json({message: 'User Deleted'})
+
   }else{
     res.status(404)
     throw new Error('User not found')
